pyproject.toml
[project]
name = "llmstrategy"
version = "0.1.0"
description = "Experiment to test end user AI models' performance on generating corporate strategies."
authors = [
    { name="Your Name", email="your@email.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "pandas",
    "matplotlib",
    "seaborn",
    "scipy",
    "pyyaml",
    "langchain",
    "langchain-openai",
    "langchain-anthropic",
    "langchain-google-genai",
    "langchain-deepseek"
]

[project.scripts]
analyze = "llmstrategy.analyze:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
include = [
    "llmstrategy/",
    "data/",
    "README.md"
]

[tool.hatch.build.targets.wheel]
include = [
    "llmstrategy/",
    "data/"
]

[tool.hatch.build]
include = [
    "data/**"
]